--- a/src/prime_rl/trainer/model.py
+++ b/src/prime_rl/trainer/model.py
@@ -191,6 +191,7 @@ def apply_ac(model: nn.Module, ac_config: ActivationCheckpointConfig):
 
 
 def apply_compile(model: nn.Module, compile_config: CompileConfig):
+    torch._dynamo.config.capture_scalar_outputs = True
     for layer_id in range(len(model.model.layers)):
         # Doing it in-place avoids mangled fqn which can break checkpoint loading
         model.model.layers[layer_id].compile(fullgraph=compile_config.fullgraph)